requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
openpyxl==3.1.5
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
